PR Feedback
Pull the compare_contracts out to its own file. This keeps the function
with its return values, but removes the noise from constants.
